+++ cocos/gui/UIListView.cpp
@@ -88,7 +88,7 @@ void ListView::updateInnerContainerSize()
     {
         case SCROLLVIEW_DIR_VERTICAL:
         {
-            int length = _items.size();
+            size_t length = _items.size();
             float totalHeight = (length - 1) * _itemsMargin;
             for (auto& item : _items)
             {
@@ -101,7 +101,7 @@ void ListView::updateInnerContainerSize()
         }
         case SCROLLVIEW_DIR_HORIZONTAL:
         {
-            int length = _items.size();
+            size_t length = _items.size();
             float totalWidth = (length - 1) * _itemsMargin;
             for (auto& item : _items)
             {
@@ -253,7 +253,7 @@ void ListView::pushBackDefaultItem()
     _refreshViewDirty = true;
 }
 
-void ListView::insertDefaultItem(int index)
+void ListView::insertDefaultItem(ssize_t index)
 {
     if (!_model)
     {
@@ -274,7 +274,7 @@ void ListView::pushBackCustomItem(Widget* item)
     _refreshViewDirty = true;
 }
 
-void ListView::insertCustomItem(Widget* item, int index)
+void ListView::insertCustomItem(Widget* item, ssize_t index)
 {
     _items.insert(index, item);
     remedyLayoutParameter(item);
@@ -282,7 +282,7 @@ void ListView::insertCustomItem(Widget* item, int index)
     _refreshViewDirty = true;
 }
 
-void ListView::removeItem(int index)
+void ListView::removeItem(ssize_t index)
 {
     Widget* item = getItem(index);
     if (!item)
@@ -305,9 +305,9 @@ void ListView::removeAllItems()
     removeAllChildren();
 }
 
-Widget* ListView::getItem(unsigned int index)
+Widget* ListView::getItem(ssize_t index)
 {
-    if ((int)index < 0 || index >= _items.size())
+    if (index < 0 || index >= _items.size())
     {
         return nullptr;
     }
@@ -319,7 +319,7 @@ Vector<Widget*>& ListView::getItems()
     return _items;
 }
 
-unsigned int ListView::getIndex(Widget *item) const
+ssize_t ListView::getIndex(Widget *item) const
 {
     if (!item)
     {
@@ -374,7 +374,7 @@ void ListView::requestRefreshView()
 
 void ListView::refreshView()
 {
-    int length = _items.size();
+    ssize_t length = _items.size();
     for (int i=0; i<length; i++)
     {
         Widget* item = _items.at(i);
@@ -427,7 +427,7 @@ void ListView::interceptTouchEvent(int handleState, Widget *sender, const Point 
     }
 }
     
-int ListView::getCurSelectedIndex() const
+ssize_t ListView::getCurSelectedIndex() const
 {
     return _curSelectedIndex;
 }
